requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==4.5.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.2
django-email-log==0.2.0
django-extensions==2.1.9
django-modalview==0.1.4
django-parsley==0.5
django-shell-plus==1.1.7
django-sslify==0.2.7
django-uuidfield==0.5.0
django-widget-tweaks==1.3
freezegun==0.3.12
funcsigs==1.0.2
gnureadline==6.3.3
gunicorn==19.3.0
importlib-metadata==1.3.0
ipdb==0.8
ipython==2.3.1
more-itertools==4.3.0
packaging==19.2
pathlib2==2.3.3
pluggy==0.8.0
psycopg2==2.7.6
py==1.7.0
pyparsing==2.4.6
pytest==4.0.1
pytest-cov==2.7.1
pytest-django==3.1.1
python-dateutil==2.8.1
pytz==2014.10
scandir==1.9.0
selenium==3.6.0
sendgrid==1.4.0
sendgrid-django==1.4.0
six==1.12.0
smtpapi==0.2.0
static3==0.6.1
typing==3.7.4
wcwidth==0.1.7
zipp==0.6.0
